Refactor all snippets and container pages to single page class
