+++ components/PhpPageContainer.php
@@ -3,15 +3,15 @@
 namespace humhub\modules\custom_pages\components;
 
 use HttpException;
-use humhub\modules\custom_pages\models\CustomContentContainer;
 use humhub\modules\custom_pages\models\forms\SettingsForm;
+use humhub\modules\custom_pages\models\Page;
 use humhub\modules\custom_pages\models\PhpType;
 use humhub\modules\file\libs\FileHelper;
 use Yii;
 use yii\helpers\Html;
 
 /**
- * @used-by CustomContentContainer
+ * @used-by Page
  */
 trait PhpPageContainer
 {
@@ -31,7 +31,7 @@ trait PhpPageContainer
             }
 
             if (!$this->validatePhpViewFile()) {
-                $this->addError($this->getPageContentProperty(), Yii::t('CustomPagesModule.base', 'Invalid view file selection!'));
+                $this->addError('page_content', Yii::t('CustomPagesModule.base', 'Invalid view file selection!'));
             }
         }
     }
@@ -45,7 +45,7 @@ trait PhpPageContainer
     public function validatePhpViewFile()
     {
         $allowedFiles = $this->getAllowedPhpViewFileSelection();
-        return array_key_exists(Html::getAttributeValue($this, $this->getPageContentProperty()), $allowedFiles);
+        return array_key_exists(Html::getAttributeValue($this, 'page_content'), $allowedFiles);
     }
 
     /**
@@ -57,7 +57,7 @@ trait PhpPageContainer
     {
         if (PhpType::isType($this->type)) {
             $viewFiles = $this->getAllowedPhpViewFileSelection(true);
-            $viewName = Html::getAttributeValue($this, $this->getPageContentProperty());
+            $viewName = Html::getAttributeValue($this, 'page_content');
 
             if (array_key_exists($viewName, $viewFiles)) {
                 return $this->getPhpViewPathByView(basename($viewFiles[$viewName]), true);
