+++ views/container/template.php
@@ -1,10 +1,12 @@
 <?php
-/** @var \humhub\modules\custom_pages\models\CustomContentContainer $page **/
+/** @var Page $page **/
 /** @var boolean $canEdit **/
 /** @var boolean $editMode **/
 /** @var string $html **/
 
+use humhub\modules\custom_pages\models\Page;
 use humhub\modules\custom_pages\modules\template\widgets\TemplatePage;
+use humhub\modules\custom_pages\modules\template\widgets\TemplatePageEditButton;
 
 $contentContainer = property_exists(Yii::$app->controller, 'contentContainer') ? Yii::$app->controller->contentContainer : null;
 ?>
@@ -12,7 +14,7 @@ $contentContainer = property_exists(Yii::$app->controller, 'contentContainer') ?
 <?php TemplatePage::begin(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode, 'contentContainer' => $contentContainer]) ?>
     <?php if($canEdit) : ?>
         <div style="margin-bottom:5px;">
-            <?= \humhub\modules\custom_pages\modules\template\widgets\TemplatePageEditButton::widget(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode]); ?>
+            <?= TemplatePageEditButton::widget(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode]); ?>
         </div>
     <?php endif; ?>
     <?= $html; ?>
