+++ views/global/edit_snippet.php
@@ -1,10 +1,11 @@
 <?php
 
 use humhub\modules\custom_pages\helpers\Url;
+use humhub\modules\custom_pages\models\Page;
 use yii\helpers\Html;
 use humhub\modules\custom_pages\modules\template\widgets\TemplatePage;
 
-/* @var $snippet \humhub\modules\custom_pages\models\CustomContentContainer */
+/* @var $snippet Page */
 /* @var $html string */
 
 $backUrl = Url::to(['/dashboard/dashboard']);
