+++ views/global/template.php
@@ -1,10 +1,12 @@
 <?php
-/** @var \humhub\modules\custom_pages\models\CustomContentContainer $page * */
-/** @var boolean $canEdit * */
-/** @var boolean $editMode * */
+/* @var Page $page */
+/* @var boolean $canEdit */
+/* @var boolean $editMode */
+/* @var string $html */
 
-/** @var string $html * */
+use humhub\modules\custom_pages\models\Page;
 use humhub\modules\custom_pages\modules\template\widgets\TemplatePage;
+use humhub\modules\custom_pages\modules\template\widgets\TemplatePageEditButton;
 ?>
 
 <?php TemplatePage::begin(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode]) ?>
@@ -12,7 +14,7 @@ use humhub\modules\custom_pages\modules\template\widgets\TemplatePage;
     <div class="col-md-12">
         <?php if ($canEdit) : ?>
             <div style="margin-bottom:5px;">
-                <?= \humhub\modules\custom_pages\modules\template\widgets\TemplatePageEditButton::widget(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode]); ?>
+                <?= TemplatePageEditButton::widget(['page' => $page, 'canEdit' => $canEdit, 'editMode' => $editMode]); ?>
             </div>
         <?php endif; ?>
         <?= $html; ?>
