+++ widgets/views/snippet_html.php
@@ -5,9 +5,10 @@
  * @license https://www.humhub.com/licences
  */
 
+use humhub\modules\custom_pages\models\Page;
 use humhub\modules\custom_pages\widgets\SnippetContent;
 
-/* @var $model \humhub\modules\custom_pages\models\CustomContentContainer */
+/* @var $model Page */
 /* @var $canEdit bool */
 
 $navigation = (!$canEdit) ? [] : [
